requirements.txt
attrs==24.2.0
beautifulsoup4==4.12.3
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
dnspython==2.6.1
exceptiongroup==1.2.2
h11==0.14.0
idna==3.7
lxml==5.2.2
outcome==1.3.0.post0
packaging==24.1
pycparser==2.22
pymongo==4.8.0
PySocks==1.7.1
python-dotenv==1.0.1
requests==2.32.3
selenium==4.23.1
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
trio==0.26.2
trio-websocket==0.11.1
typing_extensions==4.12.2
urllib3==2.2.2
webdriver-manager==4.0.2
websocket-client==1.8.0
wsproto==1.2.0
